Insert missing section title 'Use Cases With API 1'
I also changed the title of the section 'Requirements' to
'Use Cases (Requirements') to indicate that the examples of use of API 1
refer to the use cases defined in the Requirements section.

I will then update the examples in 'Use Cases with API 1' to use
the same variable names as in the section 'Use Cases (Requirements)'
where they are introduced.