Start translation of 'Use Cases with API 1'
I updated the levels and titles of headers to match the levels and titles
of headers of corresponding sections in 'Use Cases (Requirements)'.

I updated the examples to match the formulae introduced in Use Cases,
and replaced formulae of the examples with the formulae quoted from
Use Cases, copied for reference.