Add proposed API 2 ter
The difference between API 2 bis and API 2 ter is that each component
becomes a function that can select a number of its properties:
rect1("width","height") selects the "width" and "height" of component rect1.

A set of properties can then be connected to another one by calling
connect() method of the set.