Add proposed API for disconnection
I propose to return an object that represents the edge and call a method
disconnect() of this edge object to break the connection. An alternative
would be to have connect() return the disconnect() function directly.

I used a similar method in The Scalable JavaScript Application Framework
and in within.js for event listeners: the functions to create a listener
return a function to unsubscribe.